requirements.txt
pygame-ce>=2.5.3
pygame_gui>=0.6.14
